add test for integrations
