+++ src/integrations/apolloOptions.ts
@@ -11,6 +11,7 @@ import * as graphql from 'graphql';
  * - (optional) formatParams: a function applied to the parameters of every invocation of runQuery
  * - (optional) validationRules: extra validation rules applied to requests
  * - (optional) formatResponse: a function applied to each graphQL execution result
+ * - (optional) debug: a boolean that will print additional debug logging if execution errors occur
  *
  */
 interface ApolloOptions {
@@ -22,6 +23,7 @@ interface ApolloOptions {
   formatParams?: Function;
   validationRules?: Array<graphql.ValidationRule>;
   formatResponse?: Function;
+  debug?: boolean;
 }
 
 export default ApolloOptions;
